Decouple multimedia hosting from the protocol
It's probably better to let implementers host photos and videos wherever (e.g. S3) rather than require them to host them on the Sendship server.
However, it's necessary that these photos can't be changed once the post is created.
I've added a sha-256 hash attached to each multimedia content so viewers can verify this content is the same as when the post was made.